💄 Add ProjectItem component with styles and integrate into ProjectsPageV2
