+++ frontend/apps/app/features/projects/components/ProjectItem/ProjectItem.module.css
@@ -1,0 +1,106 @@
+/* Project item */
+.projectItem {
+  display: flex;
+  flex-direction: column;
+  padding: var(--spacing-5) var(--spacing-4) var(--spacing-4);
+  background-color: var(--pane-background);
+  border: 1px solid var(--global-border);
+  border-radius: var(--border-radius-md);
+  gap: var(--spacing-5);
+  transition: border-color 0.2s ease, transform 0.2s ease;
+  text-decoration: none;
+  width: 349px;
+}
+
+.projectItem:hover {
+  border-color: var(--pane-border-hover);
+  transform: translateY(-2px);
+}
+
+/* Project header */
+.projectHeader {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: var(--spacing-3);
+  width: 100%;
+}
+
+/* Project icon */
+.projectIcon {
+  width: 32px;
+  height: 32px;
+  border-radius: var(--border-radius-lg);
+  background-color: var(--global-muted-background);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  overflow: hidden;
+}
+
+.projectIconImage {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.projectIconPlaceholder {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border: 1px dashed var(--global-border);
+  color: var(--global-mute-text);
+  font-size: 16px;
+}
+
+/* Project name */
+.projectName {
+  font-size: 15px;
+  font-weight: 600;
+  color: var(--global-foreground);
+  line-height: 1.2em;
+}
+
+/* Project info */
+.projectInfo {
+  display: flex;
+  flex-direction: column;
+  gap: var(--spacing-2);
+}
+
+/* Repository badge */
+.repositoryBadge {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: var(--spacing-1);
+  padding: var(--spacing-1) var(--spacing-2) var(--spacing-1)
+    var(--spacing-1half);
+  background-color: var(--overlay-5);
+  border-radius: var(--border-radius-full);
+  width: fit-content;
+}
+
+.repositoryIcon {
+  width: 12px;
+  height: 12px;
+  color: var(--global-foreground);
+  opacity: 0.5;
+}
+
+.repositoryName {
+  font-size: 12px;
+  color: var(--global-foreground);
+}
+
+/* Commit info */
+.commitInfo {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: var(--spacing-1);
+  font-size: 11px;
+  color: var(--global-mute-text);
+}
