+++ frontend/apps/app/features/projects/components/ProjectItem/index.ts
@@ -1,0 +1,1 @@
+export { ProjectItem } from './ProjectItem'
