+++ frontend/apps/app/features/projects/components/index.ts
@@ -1,0 +1,1 @@
+export { ProjectItem } from './ProjectItem'
