+++ frontend/apps/app/features/projects/pages/ProjectsPage/ProjectsPageV2.module.css
@@ -1,10 +1,11 @@
 /* Main container */
 .container {
   display: flex;
+  width: 1548px;
+  padding: var(--spacing-20) var(--spacing-4);
   flex-direction: column;
-  gap: var(--spacing-7);
-  padding: var(--spacing-8) var(--spacing-4);
-  max-width: 1080px;
+  align-items: center;
+  gap: var(--spacing-20);
   margin: 0 auto;
 }
 
@@ -44,6 +45,7 @@
   border: 1px solid var(--global-border);
   border-radius: var(--border-radius-md);
   flex: 1;
+  height: 36px;
 }
 
 .searchInput input {
@@ -79,6 +81,7 @@
   color: var(--global-foreground);
   font-size: 13px;
   width: 282px;
+  height: 36px;
 }
 
 .sortSelectIcon {
@@ -94,15 +97,16 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  padding: var(--spacing-1) var(--spacing-3);
+  padding: var(--spacing-2) var(--spacing-3);
   background-color: var(--button-primary);
   color: var(--button-primary-foreground);
   border: 1px solid var(--button-primary);
-  border-radius: var(--border-radius-base);
+  border-radius: var(--border-radius-md);
   font-size: 13px;
   font-weight: 500;
   text-decoration: none;
   transition: background-color 0.2s ease;
+  height: 36px;
 }
 
 .newProjectButton:hover {
@@ -131,9 +135,13 @@
 
 /* Projects grid */
 .projectsGrid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(349px, 1fr));
+  display: flex;
+  height: 400px;
+  align-items: flex-start;
+  align-content: flex-start;
   gap: var(--spacing-4);
+  align-self: stretch;
+  flex-wrap: wrap;
   width: 100%;
 }
